重构: 单通道广播 + 背压 + 去低频策略 + OFI窗口化
- SPMCEventQueue: 满时先尝试清理(重读min_consumed), 仍满返回false, 调用方重试
- 单通道广播(方案A): 全部事件(成交+委托)进同一队列, 订单簿时序正确
- 去掉低频主策略 + SignalCombiner主从合成, 改为 OFI/OBI 独立信号 + 同向仲裁下单
- OFI 改滑动窗口累计(1024笔), 修复无限累计导致信号锁死不翻转
- 64位 seq: ring 头 [seq 2]->[seq 8], 彻底解决回绕
- 测试适配单通道(byte_ring_pipeline/benchmark_verify)

+++ market/parser/itch_parser.h
@@ -40,7 +40,8 @@ public:
     // 喂一条消息（不含长度前缀，len = 消息体长度）。返回是否成功处理。
     bool feed(const uint8_t* msg, size_t len);
     // 带 seq 版本：seq 是 recv 从 MoldUDP64 包头 + 包内偏移推算的消息序号。
-    bool feed(const uint8_t* msg, size_t len, uint16_t seq);
+    // uint64_t: 完整消息序号不回绕(之前 uint16_t 在 873 万条下回绕 133 次)。
+    bool feed(const uint8_t* msg, size_t len, uint64_t seq);
 
     uint64_t message_count() const { return msg_count_; }
 
@@ -66,5 +67,5 @@ private:
 
     Sink sink_;
     uint64_t msg_count_ = 0;
-    uint16_t cur_seq_ = 0;   // 当前消息序号（feed 带 seq 时设置）
+    uint64_t cur_seq_ = 0;   // 当前消息序号（feed 带 seq 时设置, 64 位不回绕）
 };
